+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.24.1] - 2020-09-04
+
+Note: This release requires `cloudsmith-api` >= `0.52.92`.
+
+### Added
+
+- Support for Package Usage Metrics API
+
 ## [0.24.0] - 2020-09-01
 
 Note: This release requires `cloudsmith-api` >= `0.52.79`.